requirement.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
greenlet==3.1.1
idna==3.10
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.0.1
python-telegram-bot-pagination==0.0.3
requests==2.32.3
SQLAlchemy==2.0.38
typing_extensions==4.12.2
urllib3==2.3.0